Reworked menu, fixed layout inclusion.
There may have been a USWDS difference when I moved to a local instance
(vs. a jekyll-uswds module). It had to go away to play with GH Pages,
and that may have have meant a problem with classes in the CSS for
USWDS.

(I understand the USWDS menu better now. And, the pages now inherit the
layouts correctly/better.)

Removed search. Until we find out that they want search, we'll pull
it.