Lesson lists, metadata cleaned up.
The rubric lessons were heavy on metadata from when we had Netlify. That
is now gone, and therefore the metadata is a mess. By that, I mean it
might fall out of sync with things, and therefore become random, noisy
data scattered all over the place. Further, it put information about a
single course into the rubric dimensions, when really, a course is
something that adds structure around the rubric, not that a particular
course should be woven through it.

Or, a course is a view on the rubric (the model), not something you want
intertwined.

This pulls all of the metadata out of the rubric pages, improves the
macros that build the individual course pages (using just the dimension
element of the rubric pages, which is good... that is a UID into the
rubric structure, and therefore something we *should* have to maintain
appropriately), and... that's it?

menu.html may have had a lot of work done on it. This brings it in line
with 2.9.0.